moviendo funcion de area de coders
