+++ src/controller/homeProfile-controller.js
@@ -6,6 +6,14 @@ import {
   deletePosts,
 } from '../model/posts-firestore-model.js';
 
+import {
+  readCodersDB,
+} from '../model/user-firestore-model.js';
+
+import {
+  templateCoders,
+} from '../view/templateHomeProfile.js';
+
 export const signOutUser = () => {
   signOut()
     .then(() => {
@@ -63,4 +71,18 @@ export const deletePostsOnClick = () => {
       });
     });
   }
+};
+
+export const codersArea = () => {
+  readCodersDB()
+    .then((querySnapshot) => {
+      let codersList = '';
+      const container = document.querySelector('.container-coders');
+      querySnapshot.forEach((refDoc) => {
+        const coder = refDoc.data();
+        codersList += templateCoders(coder.profilePicture, coder.names, coder.about);
+        return codersList;
+      });
+      container.innerHTML = codersList;
+    });
 };
